Trait pallet_checked_inflation::inflation::GetInflation

pub trait GetInflation<T: Config> {
    // Required method
    fn get_inflation_args(
        &self,
        eras_per_year: u32,
        current_supply: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    ) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance;
}

Getter trait for the inflation amount to be minted in each era.

Required Methods

fn get_inflation_args( &self, eras_per_year: u32, current_supply: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, ) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance

Returns the inflation amount to be minted per era.