Module pallet_checked_inflation::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

warnings 🔒

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Aliases

BalanceOf 🔒

The balance type of this pallet.

CurrentEra

The current era. Starts from 1 and is reset every year.

Halted

Whether the inflation process is halted.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NegativeImbalanceOf 🔒

The opaque token type for an imbalance. This is returned by unbalanced operations and must be dealt with.

NextEraStartingBlock

Block that the next era starts at.

YearStartIssuance

Total token supply at the very beginning of the year before any inflation has been minted.

YearlyInflationPerEra

The number of tokens minted at the beginning of every era during a year.