Crate pallet_checked_inflation

Checked Inflation Pallet

Config

Call

Pallet

Overview

This is a supporting pallet that provides the functionality for inflation. It is used to mint new tokens at the beginning of every era.

The amount of tokens minted is determined by the inflation method and its amount, and is configurable in the runtime, see the inflation module for the methods of inflation available and how their inflation amounts are calculated.

Most of the logic is implemented in the on_initialize hook, which is called at the beginning of every block.

Dispatchable Functions

set_first_year_supply - For configuring the pallet, sets the token’s YearStartIssuance to its current total issuance.

halt_unhalt_pallet - To start or stop the inflation process.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

inflation 🔒

Available inflation methods and resulting inflation amount generated per era.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_checked_inflation

Enums

InflationMethod

Inflation methods.

Traits

GetInflation

Getter trait for the inflation amount to be minted in each era.